+++ packages/theme_tailor/lib/src/generator/generator_annotation_matcher.dart
@@ -1,5 +1,5 @@
 import 'dart:async';
-import 'package:analyzer/dart/element/element2.dart';
+import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -7,17 +7,9 @@ abstract class GeneratorToBuffer<TAnnotation> extends GeneratorForAnnotation<TAn
   const GeneratorToBuffer();
 
   @override
-  String generateForAnnotatedElement(
-    Element2 element,
-    ConstantReader annotation,
-    BuildStep buildStep,
-  );
+  String generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep);
 
-  void generateToBuffer(
-    StringBuffer buffer,
-    Element2 element,
-    ConstantReader annotation,
-  );
+  void generateToBuffer(StringBuffer buffer, Element element, ConstantReader annotation);
 }
 
 abstract class GeneratorAnnotationMatcher<TAnnotation> extends GeneratorForAnnotation<TAnnotation> {
@@ -37,24 +29,18 @@ abstract class GeneratorAnnotationMatcher<TAnnotation> extends GeneratorForAnnot
     final tailorAnnotatedElements = library.classes.where(typeChecker.hasAnnotationOf);
 
     for (final element in tailorAnnotatedElements) {
-      getGeneratorFrom(element).generateToBuffer(
-        buffer,
+      getGeneratorFrom(
         element,
-        ConstantReader(typeChecker.firstAnnotationOf(element)),
-      );
+      ).generateToBuffer(buffer, element, ConstantReader(typeChecker.firstAnnotationOf(element)));
     }
 
     return buffer.toString();
   }
 
   @override
-  String generateForAnnotatedElement(
-    Element2 element,
-    ConstantReader annotation,
-    BuildStep buildStep,
-  ) {
+  String generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep) {
     return '';
   }
 
-  GeneratorToBuffer<TAnnotation> getGeneratorFrom(Element2 element);
+  GeneratorToBuffer<TAnnotation> getGeneratorFrom(Element element);
 }