+++ packages/theme_tailor/lib/src/generator/generator_for_annotated_class.dart
@@ -1,4 +1,4 @@
-import 'package:analyzer/dart/element/element2.dart';
+import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:theme_tailor/src/generator/generator_annotation_matcher.dart';
@@ -8,14 +8,10 @@ abstract class GeneratorForAnnotatedClass<TLibraryData, TAnnotationData, TData, 
   const GeneratorForAnnotatedClass();
 
   @override
-  void generateToBuffer(
-    StringBuffer buffer,
-    Element2 element,
-    ConstantReader annotation,
-  ) {
+  void generateToBuffer(StringBuffer buffer, Element element, ConstantReader annotation) {
     final classElement = ensureClassElement(element);
     final data = parseData(
-      parseLibraryData(classElement.library2, classElement),
+      parseLibraryData(classElement.library, classElement),
       parseAnnotation(annotation),
       classElement,
     );
@@ -24,27 +20,19 @@ abstract class GeneratorForAnnotatedClass<TLibraryData, TAnnotationData, TData, 
   }
 
   @override
-  String generateForAnnotatedElement(
-    Element2 element,
-    ConstantReader annotation,
-    BuildStep buildStep,
-  ) {
+  String generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep) {
     final buffer = StringBuffer();
     generateToBuffer(buffer, element, annotation);
     return buffer.toString();
   }
 
-  TLibraryData parseLibraryData(LibraryElement2 library, ClassElement2 element);
+  TLibraryData parseLibraryData(LibraryElement library, ClassElement element);
 
   TAnnotationData parseAnnotation(ConstantReader annotation);
 
-  ClassElement2 ensureClassElement(Element2 element);
+  ClassElement ensureClassElement(Element element);
 
-  TData parseData(
-    TLibraryData libraryData,
-    TAnnotationData annotationData,
-    ClassElement2 element,
-  );
+  TData parseData(TLibraryData libraryData, TAnnotationData annotationData, ClassElement element);
 
   void generateForData(StringBuffer buffer, TData data);
 }