+++ packages/theme_tailor/lib/src/util/extension/element_annotation_extension.dart
@@ -3,7 +3,7 @@ import 'package:source_gen/source_gen.dart';
 import 'package:theme_tailor/src/model/referenced_packages.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
-extension ElementAnnotationExtension on Element2 {
+extension ElementAnnotationExtension on Element {
   bool get isTailorAnnotation {
     return const TypeChecker.typeNamed(
       TailorMixin,