+++ packages/theme_tailor/lib/src/util/recursive_import_locator.dart
@@ -22,10 +22,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import 'package:analyzer/dart/element/element2.dart';
+import 'package:analyzer/dart/element/element.dart';
 import 'package:collection/collection.dart';
 
-extension FindAllAvailableTopLevelElements on LibraryElement2 {
+extension FindAllAvailableTopLevelElements on LibraryElement {
   bool isFromPackage(String packageName) {
     return firstFragment.source.fullName.startsWith('/$packageName/');
   }
@@ -35,28 +35,24 @@ extension FindAllAvailableTopLevelElements on LibraryElement2 {
   ///
   /// This function does not guarantees that the elements returned are unique.
   /// It is possible for the same object to be present multiple times in the list.
-  Iterable<Element2> findAllAvailableTopLevelElements() {
+  Iterable<Element> findAllAvailableTopLevelElements() {
     return _findAllAvailableTopLevelElements(
       {},
       checkExports: false,
-      key: _LibraryKey(
-        hideStatements: {},
-        showStatements: {},
-        librarySource: firstFragment.source.fullName,
-      ),
+      key: _LibraryKey(hideStatements: {}, showStatements: {}, librarySource: firstFragment.source.fullName),
     );
   }
 
-  Iterable<Element2> _findAllAvailableTopLevelElements(
+  Iterable<Element> _findAllAvailableTopLevelElements(
     Set<_LibraryKey> visitedLibraryPaths, {
     required bool checkExports,
     required _LibraryKey key,
   }) sync* {
-    yield* children2;
+    yield* children;
 
     final librariesToCheck = checkExports
-        ? fragments.expand((e) => e.libraryExports2).map(_LibraryDirectives.fromExport).nonNulls
-        : fragments.expand((e) => e.libraryImports2).map(_LibraryDirectives.fromImport).nonNulls;
+        ? fragments.expand((e) => e.libraryExports).map(_LibraryDirectives.fromExport).nonNulls
+        : fragments.expand((e) => e.libraryImports).map(_LibraryDirectives.fromImport).nonNulls;
 
     for (final directive in librariesToCheck) {
       if (!visitedLibraryPaths.add(directive.key)) {
@@ -64,60 +60,44 @@ extension FindAllAvailableTopLevelElements on LibraryElement2 {
       }
 
       yield* directive.library
-          ._findAllAvailableTopLevelElements(
-        visitedLibraryPaths,
-        checkExports: true,
-        key: directive.key,
-      )
+          ._findAllAvailableTopLevelElements(visitedLibraryPaths, checkExports: true, key: directive.key)
           .where((element) {
-        return (directive.showStatements.isEmpty && directive.hideStatements.isEmpty) ||
-            (directive.hideStatements.isNotEmpty && !directive.hideStatements.contains(element.name3)) ||
-            directive.showStatements.contains(element.name3);
-      });
+            return (directive.showStatements.isEmpty && directive.hideStatements.isEmpty) ||
+                (directive.hideStatements.isNotEmpty && !directive.hideStatements.contains(element.name)) ||
+                directive.showStatements.contains(element.name);
+          });
     }
   }
 }
 
 class _LibraryDirectives {
-  _LibraryDirectives({
-    required this.hideStatements,
-    required this.showStatements,
-    required this.library,
-  });
+  _LibraryDirectives({required this.hideStatements, required this.showStatements, required this.library});
 
   static _LibraryDirectives? fromExport(LibraryExport export) {
-    final library = export.exportedLibrary2;
+    final library = export.exportedLibrary;
     if (library == null) return null;
 
     final hideStatements = export.combinators.whereType<HideElementCombinator>().expand((e) => e.hiddenNames).toSet();
 
     final showStatements = export.combinators.whereType<ShowElementCombinator>().expand((e) => e.shownNames).toSet();
 
-    return _LibraryDirectives(
-      hideStatements: hideStatements,
-      showStatements: showStatements,
-      library: library,
-    );
+    return _LibraryDirectives(hideStatements: hideStatements, showStatements: showStatements, library: library);
   }
 
   static _LibraryDirectives? fromImport(LibraryImport export) {
-    final library = export.importedLibrary2;
+    final library = export.importedLibrary;
     if (library == null) return null;
 
     final hideStatements = export.combinators.whereType<HideElementCombinator>().expand((e) => e.hiddenNames).toSet();
 
     final showStatements = export.combinators.whereType<ShowElementCombinator>().expand((e) => e.shownNames).toSet();
 
-    return _LibraryDirectives(
-      hideStatements: hideStatements,
-      showStatements: showStatements,
-      library: library,
-    );
+    return _LibraryDirectives(hideStatements: hideStatements, showStatements: showStatements, library: library);
   }
 
   final Set<String> hideStatements;
   final Set<String> showStatements;
-  final LibraryElement2 library;
+  final LibraryElement library;
 
   _LibraryKey get key {
     return _LibraryKey(
@@ -139,11 +119,7 @@ class _LibraryDirectives {
 /// export 'library.dart' show B;
 /// ```
 class _LibraryKey {
-  _LibraryKey({
-    required this.hideStatements,
-    required this.showStatements,
-    required this.librarySource,
-  });
+  _LibraryKey({required this.hideStatements, required this.showStatements, required this.librarySource});
 
   final Set<String> hideStatements;
   final Set<String> showStatements;
@@ -153,22 +129,16 @@ class _LibraryKey {
   bool operator ==(Object other) {
     return other is _LibraryKey &&
         librarySource == other.librarySource &&
-        const SetEquality<String>().equals(
-          hideStatements,
-          other.hideStatements,
-        ) &&
-        const SetEquality<String>().equals(
-          showStatements,
-          other.showStatements,
-        );
+        const SetEquality<String>().equals(hideStatements, other.hideStatements) &&
+        const SetEquality<String>().equals(showStatements, other.showStatements);
   }
 
   @override
   int get hashCode => Object.hash(
-        librarySource,
-        const SetEquality<String>().hash(hideStatements),
-        const SetEquality<String>().hash(showStatements),
-      );
+    librarySource,
+    const SetEquality<String>().hash(hideStatements),
+    const SetEquality<String>().hash(showStatements),
+  );
 
   @override
   String toString() {